refactor: modal component
